+++ src/main/java/msg/practica/ro/dto/UserDTO.java
@@ -12,6 +12,8 @@ public class UserDTO{
 
     private String email;
 
+    private String token;
+
     public UserDTO() {
     }
 
@@ -44,4 +46,8 @@ public class UserDTO{
     public void setEmail(String email) {
         this.email = email;
     }
+
+    public String getToken(){return token;}
+
+    public void setToken(String token){this.token = token;}
 }
