--- src/main/java/msg/practica/ro/service/CustomUserDetailsService.java
@@ -1,26 +1,0 @@
-package msg.practica.ro.service;
-
-import msg.practica.ro.model.User;
-import msg.practica.ro.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class CustomUserDetailsService implements UserDetailsService {
-
-	@Autowired
-	private UserRepository userRepo;
-	
-	@Override
-	public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-		User user = userRepo.findByEmail(email);
-		if (user == null) {
-			throw new UsernameNotFoundException("User not found");
-		}
-		return new CustomUserDetails(user);
-	}
-
-}
