+++ src/main/java/msg/practica/ro/service/IUserService.java
@@ -2,6 +2,7 @@ package msg.practica.ro.service;
 
 import msg.practica.ro.dto.UserDTO;
 import msg.practica.ro.model.User;
+import org.springframework.security.core.userdetails.UserDetails;
 
 import java.io.IOException;
